// frontend/world-building-ui/src/proto/entities_pb.d.ts
export class AddictionDTO extends jspb.Message {
  getId(): string;
  setId(value: string): AddictionDTO;

  getBlueprintid(): string;
  setBlueprintid(value: string): AddictionDTO;

  getName(): string;
  setName(value: string): AddictionDTO;

  getMetadata(): Metadata | undefined;
  setMetadata(value?: Metadata): AddictionDTO;
  hasMetadata(): boolean;
  clearMetadata(): AddictionDTO;

  getUser(): string;
  setUser(value: string): AddictionDTO;

  getCampaign(): string;
  setCampaign(value: string): AddictionDTO;

  getWorld(): string;
  setWorld(value: string): AddictionDTO;

  getCharactersList(): Array<string>;
  setCharactersList(value: Array<string>): AddictionDTO;
  clearCharactersList(): AddictionDTO;
  addCharacters(value: string, index?: number): AddictionDTO;

  getTargetentity(): string;
  setTargetentity(value: string): AddictionDTO;

  getCreatedat(): number;
  setCreatedat(value: number): AddictionDTO;

  serializeBinary(): Uint8Array;
  toObject(includeInstance?: boolean): AddictionDTO.AsObject;
  static toObject(includeInstance: boolean, msg: AddictionDTO): AddictionDTO.AsObject;
  static serializeBinaryToWriter(message: AddictionDTO, writer: jspb.BinaryWriter): void;
  static deserializeBinary(bytes: Uint8Array): AddictionDTO;
  static deserializeBinaryFromReader(message: AddictionDTO, reader: jspb.BinaryReader): AddictionDTO;
}

export namespace AddictionDTO {
  export type AsObject = {
    id: string,
    blueprintid: string,
    name: string,
    metadata?: Metadata.AsObject,
    user: string,
    campaign: string,
    world: string,
    charactersList: Array<string>,
    targetentity: string,
    createdat: number,
  }
}

export class AddictionsDTO extends jspb.Message {
  getArrList(): Array<AddictionDTO>;
  setArrList(value: Array<AddictionDTO>): AddictionsDTO;
  clearArrList(): AddictionsDTO;
  addArr(value?: AddictionDTO, index?: number): AddictionDTO;

  serializeBinary(): Uint8Array;
  toObject(includeInstance?: boolean): AddictionsDTO.AsObject;
  static toObject(includeInstance: boolean, msg: AddictionsDTO): AddictionsDTO.AsObject;
  static serializeBinaryToWriter(message: AddictionsDTO, writer: jspb.BinaryWriter): void;
  static deserializeBinary(bytes: Uint8Array): AddictionsDTO;
  static deserializeBinaryFromReader(message: AddictionsDTO, reader: jspb.BinaryReader): AddictionsDTO;
}

export namespace AddictionsDTO {
  export type AsObject = {
    arrList: Array<AddictionDTO.AsObject>,
  }
}

export class CharacterMemoryDTO extends jspb.Message {
  getId(): string;
  setId(value: string): CharacterMemoryDTO;

  getBlueprintid(): string;
  setBlueprintid(value: string): CharacterMemoryDTO;

  getName(): string;
  setName(value: string): CharacterMemoryDTO;

  getMetadata(): Metadata | undefined;
  setMetadata(value?: Metadata): CharacterMemoryDTO;
  hasMetadata(): boolean;
  clearMetadata(): CharacterMemoryDTO;

  getUser(): string;
  setUser(value: string): CharacterMemoryDTO;

  getCampaign(): string;
  setCampaign(value: string): CharacterMemoryDTO;

  getWorld(): string;
  setWorld(value: string): CharacterMemoryDTO;

  getCharacter(): string;
  setCharacter(value: string): CharacterMemoryDTO;

  getFactstatusList(): Array<FactStatusDTO>;
  setFactstatusList(value: Array<FactStatusDTO>): CharacterMemoryDTO;
  clearFactstatusList(): CharacterMemoryDTO;
  addFactstatus(value?: FactStatusDTO, index?: number): FactStatusDTO;

  getImportance(): number;
  setImportance(value: number): CharacterMemoryDTO;

  getResistance(): number;
  setResistance(value: number): CharacterMemoryDTO;

  getAccumulator(): number;
  setAccumulator(value: number): CharacterMemoryDTO;

  getAcquiredat(): number;
  setAcquiredat(value: number): CharacterMemoryDTO;

  getLastupdatedat(): number;
  setLastupdatedat(value: number): CharacterMemoryDTO;

  getTagsList(): Array<string>;
  setTagsList(value: Array<string>): CharacterMemoryDTO;
  clearTagsList(): CharacterMemoryDTO;
  addTags(value: string, index?: number): CharacterMemoryDTO;

  getMemory(): string;
  setMemory(value: string): CharacterMemoryDTO;

  getTargetentity(): string;
  setTargetentity(value: string): CharacterMemoryDTO;

  getCreatedat(): number;
  setCreatedat(value: number): CharacterMemoryDTO;

  serializeBinary(): Uint8Array;
  toObject(includeInstance?: boolean): CharacterMemoryDTO.AsObject;
  static toObject(includeInstance: boolean, msg: CharacterMemoryDTO): CharacterMemoryDTO.AsObject;
  static serializeBinaryToWriter(message: CharacterMemoryDTO, writer: jspb.BinaryWriter): void;
  static deserializeBinary(bytes: Uint8Array): CharacterMemoryDTO;
  static deserializeBinaryFromReader(message: CharacterMemoryDTO, reader: jspb.BinaryReader): CharacterMemoryDTO;
}

export namespace CharacterMemoryDTO {
  export type AsObject = {
    id: string,
    blueprintid: string,
    name: string,
    metadata?: Metadata.AsObject,
    user: string,
    campaign: string,
    world: string,
    character: string,
    factstatusList: Array<FactStatusDTO.AsObject>,
    importance: number,
    resistance: number,
    accumulator: number,
    acquiredat: number,
    lastupdatedat: number,
    tagsList: Array<string>,
    memory: string,
    targetentity: string,
    createdat: number,
  }
}

export class CharacterMemoriesDTO extends jspb.Message {
  getArrList(): Array<CharacterMemoryDTO>;
  setArrList(value: Array<CharacterMemoryDTO>): CharacterMemoriesDTO;
  clearArrList(): CharacterMemoriesDTO;
  addArr(value?: CharacterMemoryDTO, index?: number): CharacterMemoryDTO;

  serializeBinary(): Uint8Array;
  toObject(includeInstance?: boolean): CharacterMemoriesDTO.AsObject;
  static toObject(includeInstance: boolean, msg: CharacterMemoriesDTO): CharacterMemoriesDTO.AsObject;
  static serializeBinaryToWriter(message: CharacterMemoriesDTO, writer: jspb.BinaryWriter): void;
  static deserializeBinary(bytes: Uint8Array): CharacterMemoriesDTO;
  static deserializeBinaryFromReader(message: CharacterMemoriesDTO, reader: jspb.BinaryReader): CharacterMemoriesDTO;
}

export namespace CharacterMemoriesDTO {
  export type AsObject = {
    arrList: Array<CharacterMemoryDTO.AsObject>,
  }
}
// ...
export class DiseaseDTO extends jspb.Message {
  getId(): string;
  setId(value: string): DiseaseDTO;

  getBlueprintid(): string;
  setBlueprintid(value: string): DiseaseDTO;

  getName(): string;
  setName(value: string): DiseaseDTO;

  getMetadata(): Metadata | undefined;
  setMetadata(value?: Metadata): DiseaseDTO;
  hasMetadata(): boolean;
  clearMetadata(): DiseaseDTO;

  getUser(): string;
  setUser(value: string): DiseaseDTO;

  getCampaign(): string;
  setCampaign(value: string): DiseaseDTO;

  getWorld(): string;
  setWorld(value: string): DiseaseDTO;

  getTagsList(): Array<string>;
  setTagsList(value: Array<string>): DiseaseDTO;
  clearTagsList(): DiseaseDTO;
  addTags(value: string, index?: number): DiseaseDTO;

  getDescription(): string;
  setDescription(value: string): DiseaseDTO;

  getSeverity(): DiseaseSeverityEnumDTO;
  setSeverity(value: DiseaseSeverityEnumDTO): DiseaseDTO;

  getCharactersList(): Array<string>;
  setCharactersList(value: Array<string>): DiseaseDTO;
  clearCharactersList(): DiseaseDTO;
  addCharacters(value: string, index?: number): DiseaseDTO;

  getTargetentity(): string;
  setTargetentity(value: string): DiseaseDTO;

  getCreatedat(): number;
  setCreatedat(value: number): DiseaseDTO;

  serializeBinary(): Uint8Array;
  toObject(includeInstance?: boolean): DiseaseDTO.AsObject;
  static toObject(includeInstance: boolean, msg: DiseaseDTO): DiseaseDTO.AsObject;
  static serializeBinaryToWriter(message: DiseaseDTO, writer: jspb.BinaryWriter): void;
  static deserializeBinary(bytes: Uint8Array): DiseaseDTO;
  static deserializeBinaryFromReader(message: DiseaseDTO, reader: jspb.BinaryReader): DiseaseDTO;
}

export namespace DiseaseDTO {
  export type AsObject = {
    id: string,
    blueprintid: string,
    name: string,
    metadata?: Metadata.AsObject,
    user: string,
    campaign: string,
    world: string,
    tagsList: Array<string>,
    description: string,
    severity: DiseaseSeverityEnumDTO,
    charactersList: Array<string>,
    targetentity: string,
    createdat: number,
  }
}

export class DiseasesDTO extends jspb.Message {
  getArrList(): Array<DiseaseDTO>;
  setArrList(value: Array<DiseaseDTO>): DiseasesDTO;
  clearArrList(): DiseasesDTO;
  addArr(value?: DiseaseDTO, index?: number): DiseaseDTO;

  serializeBinary(): Uint8Array;
  toObject(includeInstance?: boolean): DiseasesDTO.AsObject;
  static toObject(includeInstance: boolean, msg: DiseasesDTO): DiseasesDTO.AsObject;
  static serializeBinaryToWriter(message: DiseasesDTO, writer: jspb.BinaryWriter): void;
  static deserializeBinary(bytes: Uint8Array): DiseasesDTO;
  static deserializeBinaryFromReader(message: DiseasesDTO, reader: jspb.BinaryReader): DiseasesDTO;
}
// ...
export class CharacterProfessionDTO extends jspb.Message {
  getId(): string;
  setId(value: string): CharacterProfessionDTO;

  getBlueprintid(): string;
  setBlueprintid(value: string): CharacterProfessionDTO;

  getName(): string;
  setName(value: string): CharacterProfessionDTO;

  getMetadata(): Metadata | undefined;
  setMetadata(value?: Metadata): CharacterProfessionDTO;
  hasMetadata(): boolean;
  clearMetadata(): CharacterProfessionDTO;

  getUser(): string;
  setUser(value: string): CharacterProfessionDTO;

  getCampaign(): string;
  setCampaign(value: string): CharacterProfessionDTO;

  getWorld(): string;
  setWorld(value: string): CharacterProfessionDTO;

  getTagsList(): Array<string>;
  setTagsList(value: Array<string>): CharacterProfessionDTO;
  clearTagsList(): CharacterProfessionDTO;
  addTags(value: string, index?: number): CharacterProfessionDTO;

  getMemorypoolsList(): Array<string>;
  setMemorypoolsList(value: Array<string>): CharacterProfessionDTO;
  clearMemorypoolsList(): CharacterProfessionDTO;
  addMemorypools(value: string, index?: number): CharacterProfessionDTO;

  getTargetentity(): string;
  setTargetentity(value: string): CharacterProfessionDTO;

  getCreatedat(): number;
  setCreatedat(value: number): CharacterProfessionDTO;

  serializeBinary(): Uint8Array;
  toObject(includeInstance?: boolean): CharacterProfessionDTO.AsObject;
  static toObject(includeInstance: boolean, msg: CharacterProfessionDTO): CharacterProfessionDTO.AsObject;
  static serializeBinaryToWriter(message: CharacterProfessionDTO, writer: jspb.BinaryWriter): void;
  static deserializeBinary(bytes: Uint8Array): CharacterProfessionDTO;
  static deserializeBinaryFromReader(message: CharacterProfessionDTO, reader: jspb.BinaryReader): CharacterProfessionDTO;
}

export namespace CharacterProfessionDTO {
  export type AsObject = {
    id: string,
    blueprintid: string,
    name: string,
    metadata?: Metadata.AsObject,
    user: string,
    campaign: string,
    world: string,
    tagsList: Array<string>,
    memorypoolsList: Array<string>,
    targetentity: string,
    createdat: number,
  }
}

export class CharacterProfessionsDTO extends jspb.Message {
  getArrList(): Array<CharacterProfessionDTO>;
  setArrList(value: Array<CharacterProfessionDTO>): CharacterProfessionsDTO;
  clearArrList(): CharacterProfessionsDTO;
  addArr(value?: CharacterProfessionDTO, index?: number): CharacterProfessionDTO;

  serializeBinary(): Uint8Array;
  toObject(includeInstance?: boolean): CharacterProfessionsDTO.AsObject;
  static toObject(includeInstance: boolean, msg: CharacterProfessionsDTO): CharacterProfessionsDTO.AsObject;
  static serializeBinaryToWriter(message: CharacterProfessionsDTO, writer: jspb.BinaryWriter): void;
  static deserializeBinary(bytes: Uint8Array): CharacterProfessionsDTO;
  static deserializeBinaryFromReader(message: CharacterProfessionsDTO, reader: jspb.BinaryReader): CharacterProfessionsDTO;
}

export namespace CharacterProfessionsDTO {
  export type AsObject = {
    arrList: Array<CharacterProfessionDTO.AsObject>,
  }
}